rename: AddMail -> AddMailSection
