+++ src/components/styles/AddMailSectionStyle.jsx
